+++ src/Components/index.ts
@@ -1,3 +1,4 @@
 export {AppLayout} from './Layout/AppLayout';
 export {ListItem} from './ListItem/ListItem';
 export * from './UserCard/UserCard';
+export * from './AddCommentBox/AddCommentBox';
