+++ src/Pages/index.ts
@@ -1,7 +1,7 @@
 export {Feed} from './Feed/Feed';
 export {UserProfile} from './UserProfile/UserProfile';
 export {Connections} from './UserProfile/Connections/Connections';
-export {Tweet} from './UserProfile/Tweet/Tweet';
+export {Tweet} from './Tweet/Tweet';
 export {Login} from './Login/Login';
 export {Signup} from './Signup/Signup';
 export {Redirects} from './Redirects/Redirects';
