+++ src/app/features/bookmarkSlice.ts
@@ -1,6 +1,6 @@
-import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import axios from "axios";
-import { dislikeTweet, likeTweet } from "./tweetSlice";
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import { dislikeTweet, likeTweet } from "./thunkApiCalls/tweetThunk";
 
 export const bookmarkTweet:any = createAsyncThunk("bookmark/add",async (tweetId:string,{rejectWithValue})=>{
 	try {
@@ -36,9 +36,8 @@ export const removeBookmark:any = createAsyncThunk("bookmark/remove",async (book
 	}
 });
 
-
 export const getBookmarks:any = createAsyncThunk(
-  "user/fetchBookmark",
+  "user/getBookmark",
   async () => {
     try {
       const { data } = await axios.get("https://social-app-twitter.herokuapp.com/api/v1/user/bookmarks",{
