+++ src/app/features/index.ts
@@ -2,3 +2,4 @@ export * from './authSlice';
 export * from './userSlice';
 export * from './tweetSlice';
 export * from './bookmarkSlice';
+export * from './thunkApiCalls/tweetThunk'
